+++ src/parser/companions/types/FlamingSphere.mjs
@@ -94,13 +94,13 @@ export async function getFlamingSphere({
   stub = await DDBCompanionMixin.addEnrichedImageData(stub);
   const enriched = foundry.utils.getProperty(document, "flags.monsterMunch.enrichedImages");
 
-  console.warn("Flaming Sphere", {
-    stub: deepClone(stub),
-    enriched,
-    description,
-    bonusDescription,
-    raw,
-  });
+  // console.warn("Flaming Sphere", {
+  //   stub: deepClone(stub),
+  //   enriched,
+  //   description,
+  //   bonusDescription,
+  //   raw,
+  // });
 
   const result = {
     FlamingSphere: {